Revert "Continue migrating to Temporal"
Undo the change from OffsetDateTime to Temporal.

The Temporal API documentation states that "This interface is a
framework-level interface that should not be widely used in application code.
Instead, applications should create and pass around instances of concrete
types, such as LocalDate."

Further, the OffsetDateTime API docs state that "It is intended that
ZonedDateTime or Instant is used to model data in simpler applications.
This class may be used when modeling date-time concepts in more detail,
or when communicating to a database or in a network protocol."

For those reasons, this project will primarily use the OffsetDateTime
class in the public API from this point forward.

This reverts commit 30039132cadb76bb685d1f22ee5b218442311ad5.